+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-Fix: IE11 Button height compatibility fix
+- Fix: IE11 Button height compatibility fix
 
 ## [2.6.0] - 2019-05-15