+++ CHANGELOG.md
@@ -4,7 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Expose `setSticky`/`setUnsticky` methods in StickyContainer.
+## [2.8.0] - 2019-06-03
+
+Feature:
+- Expose `setSticky`/`setUnsticky` methods in StickyContainer.
 
 ## [2.7.0] - 2019-05-24